refactor (move) project state persister
